renamed plugin dir to plugins because that's the proper name; added MixinPlugin to the other top-level extensions
